New Moon | Reinforced Hoods | Charring Bladeclass
OldRW classic that inspired hate and disdain for Tabaxi everywhere, now once more returned. We've got Forlorn, which is Lupian/Wild-Kin exclusive. Grudgebearers, as Dwarf exclusive. Black Oak, as Elf exclusive. Oathmarked, as Drakian exclusive. Etc. You get the idea. Now, we've got New Moon, which is Tabaxi/Wild-Kin exclusive.

If you're familiar with OldRW NM, it's just that, but with expert swords, 15 spellpoints, mage armour, summon weapon, frost bolt and air slash. That's about it. Seriously. Almah but Tabaxi, effectively, with a unique spell(kind of) and the funny sword.

On that note, the sword has been tweaked / updated, with an objective nerf to force, when on, since burn damage is nasty. I've considered removing the burn damage and just giving it all the bladeclass below. I'll probably do that before this is fullmerged, if we even humour it.

Charring bladeclass introduced, PROPERLY, given wounds that have a high healthpool, high pain. They're exclusive to this weapon for now. I've other ideas to introduce it. Sizzle sounds from elsewhere introduced. I don't really know who/where to credit for them.

Additionally, reinforced hoods have been reintroduced from OldRW. Given NM had them, and I've had players ask repeatedly for 'em. Their armour values were rounded up. The item was added to the seamster's vendor, and it's now available as an expensive, late game craft. Given it can be used for armour layering and it's super good as a result. The only source of it outside of that vendor and crafting, as of now, are NM mercs.