Rituos Return | Rituos Rite Removal
Simply returns old functionality of rituos, where you sacrificed one body part to be skeletonized a day, per day, until it was just your head left, in order to gain the rite's value of spellpoints and boons.

The rite let you skip this process - in fact, it was removed from rituos altogether - for some awful reason. Frame one rituos lich on spawn fragfests were not great. There was no longevity or buildup for power. It was immediate and cringe.

AP had the right idea with how this functioned previously. If we want the ritual circle stuff back, we can return it in a better light. But that wasn't it.